refactor(Deps.ts): Relocate deps.ts and remove the duplicates
