--- lib/deps.ts
@@ -1,17 +1,0 @@
-/**
- * Application dependencies
- */
-
-export { join } from "https://deno.land/std@0.177.0/path/mod.ts";
-export { parse } from "https://deno.land/std@0.177.0/flags/mod.ts";
-export * as logger from "https://deno.land/std@0.177.0/log/mod.ts";
-export { z } from "https://deno.land/x/zod@v3.16.1/mod.ts";
-export {
-  parse as parseYAML,
-  stringify,
-} from "https://deno.land/std@0.177.0/encoding/yaml.ts";
-export { crypto } from "https://deno.land/std@0.177.0/crypto/crypto.ts";
-export {
-  copySync,
-  emptyDirSync,
-} from "https://deno.land/std@0.177.0/fs/mod.ts";
