+++ test/note.test.ts
@@ -1,5 +1,5 @@
 import { Frontmatter, Note } from "../lib/note.ts";
-import { assertEquals, beforeEach, describe, it } from "./deps.ts";
+import { assertEquals, beforeEach, describe, it } from "../deps.ts";
 
 import { Emitter } from "../lib/eventEmitter.ts";
 import { LinkManager } from "../lib/linkManager.ts";
